refactor: 각 모듈에서 상수 불러올 때 한 곳에 모아둔 곳에서부터 불러오게해라
