+++ src/app/asset/util/calcCoinAmount.js
@@ -1,4 +1,4 @@
-import COIN from '../constants/COIN.js';
+import { COIN } from '../constants/index.js';
 
 const calcCoinAmount = (coinCnts) =>
     coinCnts.reduce((m, coinCnt, idx) => m + COIN[idx] * coinCnt, 0);
