+++ src/app/asset/util/getEnableChargeAmount.js
@@ -1,4 +1,4 @@
-import INPUT_NUMBER_RULE from '../constants/INPUT_NUMBER_RULE.js';
+import { INPUT_NUMBER_RULE } from '../constants/index.js';
 
 const getEnableChargeAmount = (chargeAmount) =>
     chargeAmount - (chargeAmount % INPUT_NUMBER_RULE.mod);
