+++ src/app/asset/validation/checkProductQuantity.js
@@ -1,5 +1,5 @@
 import { createCheckNumber } from './index.js';
-import ERROR_MSG from '../constants/ERROR_MSG.js';
+import { ERROR_MSG } from '../constants/index.js';
 
 const checkProductQuantity = createCheckNumber(ERROR_MSG.requireProductQuantity);
