+++ src/app/localStorage/inputAmount.js
@@ -1,4 +1,4 @@
-import STORAGE_KEY from '../asset/constants/STORAGE_KEY.js';
+import { STORAGE_KEY } from '../asset/constants/index.js';
 import { getNumber } from './index.js';
 
 export const getInputAmount = () => getNumber(STORAGE_KEY.inputAmount);
